simplify lutrap routing
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@5745 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lind/os-jos64/drivers/jif_user.c
@@ -4,6 +4,7 @@
 #include <inc/assert.h>
 #include <inc/error.h>
 #include <inc/jthread.h>
+#include <inc/intmacro.h>
 #include <machine/memlayout.h>
 
 #include <string.h>
@@ -115,7 +116,7 @@ jif_rx_thread(void *a)
 	jif->irq_flag = 1;
 	lutrap_kill(SIGNAL_ETH);
 	while (jif->irq_flag)
-	    sys_sync_wait(&jif->irq_flag, 1, ~0UL);
+	    sys_sync_wait(&jif->irq_flag, 1, UINT64(~0UL));
     }
 }
